# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Collaboration Platform API
  version: 0.0.1-SNAPSHOT
  description: REST API contract for the modular collaboration platform backend.
servers:
  - url: http://localhost:8080
    description: Local backend
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/workspaces/{workspaceId}/events:
    get:
      tags:
        - Realtime
      summary: Poll versioned realtime events for a workspace.
      description: Polling baseline for realtime comparison. Use `nextCursor` as the next `since` value.
      operationId: pollRealtimeEventsV1
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/SinceCursor"
      responses:
        "200":
          description: Versioned workspace events after the supplied cursor.
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealtimePollResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v2/workspaces/{workspaceId}/events/stream:
    get:
      tags:
        - Realtime
      summary: Stream versioned realtime events for a workspace using Server-Sent Events.
      description: SSE comparison endpoint. Each event uses the event version as the SSE `id`.
      operationId: streamRealtimeEventsV2
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/SinceCursor"
      responses:
        "200":
          description: Server-Sent Event stream of versioned realtime events.
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/RealtimeEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session-token
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
      description: Optional client-supplied correlation ID echoed by the backend.
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
...
    SinceCursor:
      name: since
      in: query
      required: false
      schema:
        type: integer
        format: int64
        default: 0
        minimum: 0
      description: Last observed realtime event version. Events with greater versions are returned.
  headers:
    CorrelationId:
      description: Correlation ID assigned to the request.
      schema:
        type: string
  responses:
...
    Unauthorized:
      description: Missing, malformed, expired, or revoked bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated actor is not allowed to perform the operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    RealtimePollResponse:
      type: object
      required:
        - nextCursor
        - events
      properties:
        nextCursor:
          type: integer
          format: int64
        events:
          type: array
          items:
            $ref: "#/components/schemas/RealtimeEvent"
    RealtimeEvent:
      type: object
      required:
        - version
        - schemaVersion
        - eventId
        - eventType
        - tenantId
        - aggregateId
        - occurredAt
        - recordedAt
        - payload
      properties:
        version:
          type: integer
          format: int64
        schemaVersion:
          type: string
          example: realtime.event.v1
        eventId:
          type: string
        eventType:
          type: string
          example: DocumentCreatedEvent
        tenantId:
          type: string
        workspaceId:
          type:
            - string
            - "null"
        aggregateId:
          type: string
        occurredAt:
          type: string
          format: date-time
        recordedAt:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - BAD_REQUEST
            - CONFLICT
            - UNAUTHORIZED
            - FORBIDDEN
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...